Remove socket.io-encrypt
This module has no license information which makes it hard to use in
enterprise.  It also doesnt appear to be supported any longer and peer
depends on a very old version of socket.io

Also cleaned up some of the options and added a url option so you can directly
set what gets passed to the socket client if you choose.